Started working on getting Schematic PSR-4 compatible
