+++ src/models/Result.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace NerdsAndCompany\Schematic\Models;
+
+use Craft\BaseModel as Base;
+
+/**
+ * Schematic Result Model.
+ *
+ * Encapsulates the result of an action, including error messages.
+ *
+ * @author    Nerds & Company
+ * @copyright Copyright (c) 2015, Nerds & Company
+ * @license   MIT
+ *
+ * @link      http://www.nerds.company
+ */
+class Result extends Base
+{
+    /**
+     * Consumes the errors listed in an existing result and appends them to this result.
+     *
+     * @param Schematic_ResultModel $result The result to consume.
+     */
+    public function consume(Result $result)
+    {
+        if ($result->hasErrors('errors')) {
+            $this->addErrors(array('errors' => $result->getErrors('errors')));
+        }
+    }
+}
